requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu
pandas
python-dotenv
streamlit